requirements.txt
absl-py==1.4.0
alembic==1.10.2
astroid==2.15.2
Box2D==2.3.10
cachetools==5.3.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
chardet==5.1.0
charset-normalizer==3.1.0
cloudpickle==2.2.1
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
coverage==7.2.2
cycler==0.11.0
DataProperty==0.55.0
dill==0.3.6
docopt==0.6.2
Farama-Notifications==0.0.4
filelock==3.10.7
fonttools==4.39.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.53.0
gym==0.21.0
gymnasium==0.28.1
huggingface-hub==0.13.3
huggingface-sb3==2.2.4
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
isort==5.12.0
jax-jumpy==1.0.0
Jinja2==3.1.2
-e git+ssh://git@github.com/maxrudolph1/fuzzrl.git@67c8bb1b5b0cafb74f4604f529b0ef4e8cce9809#egg=jpeg_encode&subdirectory=jpeg_encode
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mbstrdecoder==1.1.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
optuna==3.1.0
packaging==23.0
pandas==2.0.0
pathvalidate==2.5.2
Pillow==9.5.0
pipreqs==0.4.11
platformdirs==3.2.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pylint==2.17.2
pylint-runner==0.6.0
pyparsing==3.0.9
pytablewriter==0.64.2
python-dateutil==2.8.2
PyTurboJPEG==1.7.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
-e git+https://github.com/DLR-RM/rl-baselines3-zoo@8729ec224e47890f630ae35c113ed5660e49b259#egg=rl_zoo3
rsa==4.9
sb3-contrib==1.8.0a13
scipy==1.10.1
Shimmy==0.2.1
simplejpeg==1.6.5
six==1.16.0
SQLAlchemy==2.0.8
stable-baselines3==2.0.0a1
sympy==1.11.1
tabledata==1.3.1
tcolorpy==0.1.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
tomlkit==0.11.7
torch==2.0.0
tqdm==4.65.0
typepy==1.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0
